added user implementation

+++ internal/storage/postgresql/postgresql.go
@@ -16,7 +16,7 @@ type Storage struct {
 	db *pgxpool.Pool
 }
 
-const returning = "RETURNING id, title, description, priority, COALESCE(status, '') as status, created_at, due_date"
+const returning = "RETURNING id, user_id, title, description, priority, COALESCE(status, '') as status, created_at, due_date"
 
 func New(storagePath string) (*Storage, error) {
 	const op = "storage.postgresql.New"
@@ -27,12 +27,12 @@ func New(storagePath string) (*Storage, error) {
 	return &Storage{db: db}, nil
 }
 
-func (s *Storage) CreateTask(ctx context.Context, title, description string,
+func (s *Storage) CreateTask(ctx context.Context, uid uint64, title, description string,
 	priority string) (*models.Task, error) {
 	const op = "storage.postgresql.CreateTask"
 	var task models.Task
-	err := pgxscan.Get(ctx, s.db, &task, "INSERT INTO tasks(title, description, priority) "+
-		"VALUES ($1, $2, $3)"+returning, title, description, priority)
+	err := pgxscan.Get(ctx, s.db, &task, "INSERT INTO tasks(user_id, title, description, priority) "+
+		"VALUES ($1, $2, $3, $4)"+returning, uid, title, description, priority)
 	if err != nil {
 		if lerr := checkTooLongField(op, err); lerr != nil {
 			return nil, lerr
@@ -42,13 +42,13 @@ func (s *Storage) CreateTask(ctx context.Context, title, description string,
 	return &task, nil
 }
 
-func (s *Storage) GetTask(ctx context.Context, id uint64) (*models.Task, error) {
+func (s *Storage) GetTask(ctx context.Context, id uint64, uid uint64) (*models.Task, error) {
 	const op = "storage.postgresql.GetTask"
 	var task models.Task
 	err := pgxscan.Get(ctx, s.db, &task, ""+
-		"SELECT id, title, description, priority, "+
+		"SELECT id, user_id, title, description, priority, "+
 		"COALESCE(status, '') as status, created_at, "+
-		"due_date FROM tasks WHERE id = $1", id)
+		"due_date FROM tasks WHERE id = $1 AND user_id = $2", id, uid)
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) {
 			return nil, fmt.Errorf("%s: %w", op, storage.ErrTaskNotFound)
@@ -58,18 +58,19 @@ func (s *Storage) GetTask(ctx context.Context, id uint64) (*models.Task, error) 
 	return &task, nil
 }
 
-func (s *Storage) UpdateTask(ctx context.Context, id uint64, title, description, priority string) (*models.Task, error) {
+func (s *Storage) UpdateTask(ctx context.Context, id uint64, uid uint64, title, description string,
+	priority string) (*models.Task, error) {
 	const op = "storage.postgresql.GetTask"
 	var task models.Task
 	query := `
         UPDATE tasks 
         SET 
-            title = COALESCE(NULLIF($2, ''), title),
-            description = COALESCE(NULLIF($3, ''), description),
-            priority = COALESCE(NULLIF($4, ''), priority)
-        WHERE id = $1 
+            title = COALESCE(NULLIF($3, ''), title),
+            description = COALESCE(NULLIF($4, ''), description),
+            priority = COALESCE(NULLIF($5, ''), priority)
+        WHERE id = $1 AND user_id = $2
     ` + returning
-	err := pgxscan.Get(ctx, s.db, &task, query, id, title, description, priority)
+	err := pgxscan.Get(ctx, s.db, &task, query, id, uid, title, description, priority)
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) {
 			return nil, fmt.Errorf("%s: %w", op, storage.ErrTaskNotFound)
@@ -82,10 +83,10 @@ func (s *Storage) UpdateTask(ctx context.Context, id uint64, title, description,
 	return &task, nil
 }
 
-func (s *Storage) UpdateStatus(ctx context.Context, id uint64, status string) (*models.Task, error) {
+func (s *Storage) UpdateStatus(ctx context.Context, id uint64, uid uint64, status string) (*models.Task, error) {
 	const op = "storage.postgresql.UpdateStatus"
 	var task models.Task
-	err := pgxscan.Get(ctx, s.db, &task, "UPDATE tasks SET status = $1 WHERE id = $2 RETURNING *", status, id)
+	err := pgxscan.Get(ctx, s.db, &task, "UPDATE tasks SET status = $1 WHERE id = $2 AND user_id = $3 RETURNING *", status, id, uid)
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) {
 			return nil, fmt.Errorf("%s: %w", op, storage.ErrTaskNotFound)
@@ -95,9 +96,9 @@ func (s *Storage) UpdateStatus(ctx context.Context, id uint64, status string) (*
 	return &task, nil
 }
 
-func (s *Storage) DeleteTask(ctx context.Context, id uint64) error {
+func (s *Storage) DeleteTask(ctx context.Context, id uint64, uid uint64) error {
 	const op = "storage.postgresql.DeleteTask"
-	commandTag, err := s.db.Exec(ctx, "DELETE FROM tasks WHERE id = $1", id)
+	commandTag, err := s.db.Exec(ctx, "DELETE FROM tasks WHERE id = $1 AND user_id = $2", id, uid)
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) {
 			return fmt.Errorf("%s: %w", op, storage.ErrTaskNotFound)
